+++ backend/src/main/java/org/usyj/makgora/response/voteDetails/VoteDetailSettlementResponse.java
@@ -21,8 +21,8 @@ public class VoteDetailSettlementResponse {
     @Setter
     @Builder
     public static class OptionSettlementResult {
-        private Integer optionId;
-        private Integer correctChoiceId;
+        private Long optionId;
+        private Long correctChoiceId;
         private Double odds;
         private int optionPool;
         private int winnerPool;
